update index toggle chat
